+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- log timestamp has precision in milliseconds instead of seconds.
+
 ## v3.2.0 - 2020-11-13
 
 ### Added